feat: use Goodreads description when available (#73)
This PR corresponds with chrisvogt/gatsby-theme-chrisvogt#359. I'm
unhappy with the descriptions coming back from the Google Books API.
They do not match what's rendered on Google Books, and in the case of
_Flowers for Algernon_, I find the description short, crude, and
offensive—not something I want rendered on my personal website.

This PR updates the API to use the Goodreads description when that's
available, and preserves using the other data from Google Books.